Create a Rewinder class to store state rewinding logic
