+++ Assets/Plugins/Editor/Uplift/DependencyResolution/State.cs
@@ -170,8 +170,8 @@ namespace Uplift.DependencyResolution
 	{
 		DependencyState parent;
 		public List<PossibilitySet> matchingPossibilitySet = new List<PossibilitySet>();
-		DependencyDefinition currentRequirement;
-		List<Conflict> unusedUnwinds = new List<Conflict>(); //FIXME Unwind = conflict + stateToUnwind
+		public DependencyDefinition currentRequirement;
+		List<Conflict> unusedUnwinds = new List<Conflict>();
 
 		public PossibilityState(string name,
 								Stack<DependencyDefinition> requirements,
@@ -292,187 +292,6 @@ namespace Uplift.DependencyResolution
 			Conflict newConflict = new Conflict(currentRequirement, activated);
 			conflicts.Add(newConflict);
 			Debug.Log(newConflict);
-		}
-
-		private List<string> FindConflictingDependency(Conflict conflictRaised)
-		{
-			//FIXME conflict might be due to several requirement
-			List<string> conflictingDependencies = new List<string>();
-			DependencyNode correspondingNode = conflictRaised.activated.FindByName(conflictRaised.requirement.Name);
-
-			DependencyNode copiedNode = new DependencyNode();
-			copiedNode.restrictions = correspondingNode.restrictions;
-			if (copiedNode.restrictions.Keys != null && copiedNode.restrictions.Keys.Count > 0)
-			{
-				List<String> restrictors = new List<string>(copiedNode.restrictions.Keys);
-
-				foreach (string key in restrictors)
-				{
-					//Initial requirements cannot be changed
-					if (key == "initial")
-					{
-						continue;
-					}
-					//TODO create a function to do this
-					IVersionRequirement tmpVersion = copiedNode.restrictions[key];
-					copiedNode.restrictions[key] = new NoRequirement();
-					try
-					{
-						IVersionRequirement newRequirementForNode = copiedNode.ComputeRequirement();
-						//IVersionRequirement versionRequirementTobeChecked = conflictRaised.requirement.Requirement.RestrictTo(newRequirementForNode);
-
-						foreach (Upset package in correspondingNode.selectedPossibilitySet.packages)
-						{
-							if (newRequirementForNode.IsMetBy(package.PackageVersion))
-							{
-								Debug.Log("When removing " + key + " requirements on " + conflictRaised.requirement.Name + " no conflict remains.");
-								conflictingDependencies.Add(key);
-								break;
-							}
-						}
-					}
-					catch (Exception e)
-					{
-						Debug.LogError(e.ToString());
-					}
-					finally
-					{
-						copiedNode.restrictions[key] = tmpVersion;
-					}
-				}
-			}
-
-			return conflictingDependencies;
-		}
-
-		//TODO Move rewind logic in a separated class
-		public Stack<State> UnwindForConflict(Stack<State> stack) //return stack ?
-		{
-			Conflict conflict = conflicts.ToArray()[0];
-			//In dependencyState ?
-			Debug.Log("unwind for conflict started");
-			// Find requirements causing conflict
-			Debug.Log("Get list of conflicting requirements");
-
-			//TODO if null raise error ?
-			List<string> conflictingRequirements = FindConflictingDependency(conflict);
-
-			Stack<State> stackCopy = new Stack<State>(stack);
-			List<State> possibleRewinds = new List<State>();
-
-			Debug.Log("Go through possible rewinds");
-			foreach (string nodeName in conflictingRequirements)
-			{
-				possibleRewinds.AddRange(GetPossibleRewinds(stackCopy, nodeName));
-
-				//TODO look for unwinds in previous unused unwinds
-				// --> Check unwind that were not executed (in list of unused unwinds)
-				// * Keep only those that are smaller than those in the array[]
-				// * Check if unwind has the chance of prevent encounter current conflict
-				//		* the unwind must have been rejected an unwind leading to
-				//		* one of the state in the current conflict reqt tree
-
-				// --> Launch unwind
-				// * If unwinds are found take the smallest one
-				//		* filter destination state's possibilities to prevent conflict
-				// * else
-				//		* raise VersionConflict error
-				// -->  Update list of unused unwinds			
-			}
-
-			PossibilityState rewindStateCandidate = FindStateWithGreaterIndex(possibleRewinds);
-			Debug.Log("Rewind candidate is found.");
-			Debug.Log(rewindStateCandidate.currentRequirement.Name + " has other possibility sets matching requirements");
-			//FIXME Remove it from possible rewind and put the remaining one in unusedUnwinds
-
-			int depthToRewind = rewindStateCandidate.depth;
-			conflicts.Remove(conflict);
-			RewindToState(stack, depthToRewind);
-
-			return stack;
-		}
-
-		private Stack<State> RewindToState(Stack<State> stack, int depthToRewind)
-		{
-			Debug.Log("rewinding to state at depth " + depthToRewind);
-			while (stack.Peek().depth > depthToRewind)
-			{
-				stack.Pop();
-			}
-			//Use other possibilitySet than selected possibilitySet
-			PossibilityState stateToRewind = (PossibilityState)stack.Pop();
-			DependencyNode correspondingNode = stateToRewind.activated.FindByName(stateToRewind.currentRequirement.Name);
-
-			Debug.Log("removing previously selected possibility set");
-			PossibilitySet selectedPossibilitySet = correspondingNode.selectedPossibilitySet;
-			//FIXME : super dangerous to have duplicate matching possibilitySet...
-			stateToRewind.matchingPossibilitySet.Remove(selectedPossibilitySet);
-
-			Debug.Log("Remove it from matching possibility set");
-			correspondingNode.matchingPossibilities = stateToRewind.matchingPossibilitySet;
-
-			Debug.Log("Push edited possibility state to stack");
-			stack.Push(stateToRewind);
-			return stack;
-		}
-
-		private List<State> GetPossibleRewinds(Stack<State> stack, string dependencyName)
-		{
-			State currentState = stack.Pop();
-
-			List<State> matchingPossibilityStates = new List<State>();
-			List<State> listOfStates = new List<State>(stack.ToArray());
-			List<State> listOfPossibleRewinds = new List<State>();
-
-			//_Check parent of dependencyName for alternatives, check parent alternative, check gramps alternatives
-			//Do it recursively until top
-			//Add possibilityState of dependencyName
-			if (stack.Count >= 1)
-			{
-				//TODO Use linq !!!
-				listOfPossibleRewinds.Add(listOfStates.Find(state => state.GetType() == typeof(PossibilityState)
-																&& ((PossibilityState)state).currentRequirement.Name == dependencyName
-																&& state.activated.FindByName(dependencyName).matchingPossibilities.Count > 1));
-
-				foreach (DependencyNode node in currentState.activated.FindNodesWithGivenDependency(dependencyName))
-				{
-					//TODO linq
-					matchingPossibilityStates = listOfStates.FindAll(state => state.GetType() == typeof(PossibilityState)
-																		&& ((PossibilityState)state).currentRequirement.Name == node.Name
-																		&& state.activated.FindByName(node.Name).matchingPossibilities.Count > 1);
-					listOfPossibleRewinds.AddRange(matchingPossibilityStates);
-				}
-			}
-			return listOfPossibleRewinds;
-
-			//Explore parent and check if they have alternative possibility sets with differents dependencies
-			/*
-			if (currentState.GetType() == typeof(PossibilityState)
-			&& (currentState.possibilities.Count > 1))
-			{
-				// ;
-				// Get nodes.name and look for corresponding possibility states
-				List<PossibilityState> result = GetPossibleRewinds(stack, currentState.name);
-				result.Add((PossibilityState)currentState);
-				return result; //FIXME just add depth of state is better ?
-			}
-			else
-			{
-				return GetPossibleRewinds(stack, dependencyName);
-			}
-			*/
-		}
-		private PossibilityState FindStateWithGreaterIndex(List<State> possibleRewinds)
-		{
-			PossibilityState rewindCandidate = null;
-			foreach (PossibilityState state in possibleRewinds)
-			{
-				if (rewindCandidate == null || rewindCandidate.depth < state.depth)
-				{
-					rewindCandidate = state;
-				}
-			}
-			return rewindCandidate;
 		}
 	}
 }
